Split the asyncio module
Since more code is coming, and this module is already a bit large, we
split it into a util and a service sub-module.
